feat:updated database routes for pushing to data, still need to writeup database routes for pulling from database
